[Feat]: base_dataset.py 제작 - SeqDataset, NonSeqDataset 간소화 및 오류 수정하여 추가 - dataset.py에서 이들을 상속하여 사용
